refactor(backend): add payments subfolder
