+++ packages/backend/src/app.ts
@@ -26,9 +26,9 @@ import { PeerService } from './peer/service'
 import { AccountService } from './open_payments/account/service'
 import { RatesService } from './rates/service'
 import { SPSPRoutes } from './spsp/routes'
-import { IncomingPaymentRoutes } from './open_payments/incoming/routes'
+import { IncomingPaymentRoutes } from './open_payments/payments/incoming/routes'
 import { AccountRoutes } from './open_payments/account/routes'
-import { IncomingPaymentService } from './open_payments/incoming/service'
+import { IncomingPaymentService } from './open_payments/payments/incoming/service'
 import { StreamServer } from '@interledger/stream-receiver'
 import { WebhookService } from './webhook/service'
 import { OutgoingPaymentService } from './outgoing_payment/service'
