+++ packages/backend/src/open_payments/payments/incoming/routes.test.ts
@@ -4,22 +4,22 @@ import Knex from 'knex'
 import { WorkerUtils, makeWorkerUtils } from 'graphile-worker'
 import { v4 as uuid } from 'uuid'
 
-import { createContext } from '../../tests/context'
-import { AccountService } from '../account/service'
-import { Account } from '../account/model'
-import { createTestApp, TestContainer } from '../../tests/app'
-import { resetGraphileDb } from '../../tests/graphileDb'
-import { GraphileProducer } from '../../messaging/graphileProducer'
-import { Config, IAppConfig } from '../../config/app'
+import { createContext } from '../../../tests/context'
+import { AccountService } from '../../account/service'
+import { Account } from '../../account/model'
+import { createTestApp, TestContainer } from '../../../tests/app'
+import { resetGraphileDb } from '../../../tests/graphileDb'
+import { GraphileProducer } from '../../../messaging/graphileProducer'
+import { Config, IAppConfig } from '../../../config/app'
 import { IocContract } from '@adonisjs/fold'
-import { initIocContainer } from '../..'
-import { AppServices } from '../../app'
-import { truncateTables } from '../../tests/tableManager'
-import { randomAsset } from '../../tests/asset'
+import { initIocContainer } from '../../..'
+import { AppServices } from '../../../app'
+import { truncateTables } from '../../../tests/tableManager'
+import { randomAsset } from '../../../tests/asset'
 import { IncomingPaymentService } from './service'
 import { IncomingPayment } from './model'
 import { IncomingPaymentRoutes, MAX_EXPIRY } from './routes'
-import { AppContext } from '../../app'
+import { AppContext } from '../../../app'
 
 describe('Incoming Payment Routes', (): void => {
   let deps: IocContract<AppServices>
