+++ packages/backend/src/open_payments/payments/incoming/routes.ts
@@ -1,10 +1,10 @@
 import base64url from 'base64url'
 import { StreamServer } from '@interledger/stream-receiver'
 import { Logger } from 'pino'
-import { validateId } from '../../shared/utils'
-import { AppContext } from '../../app'
-import { IAppConfig } from '../../config/app'
-import { AccountingService } from '../../accounting/service'
+import { validateId } from '../../../shared/utils'
+import { AppContext } from '../../../app'
+import { IAppConfig } from '../../../config/app'
+import { AccountingService } from '../../../accounting/service'
 import { IncomingPaymentService } from './service'
 import { IncomingPayment } from './model'
