+++ packages/backend/src/outgoing_payment/service.test.ts
@@ -24,7 +24,7 @@ import { RETRY_BACKOFF_SECONDS } from './worker'
 import { isTransferError } from '../accounting/errors'
 import { AccountingService, TransferOptions } from '../accounting/service'
 import { AssetOptions } from '../asset/service'
-import { IncomingPayment } from '../open_payments/incoming/model'
+import { IncomingPayment } from '../open_payments/payments/incoming/model'
 import { RatesService } from '../rates/service'
 import { Pagination } from '../shared/baseModel'
 import { getPageTests } from '../shared/baseModel.test'
